+++ clang-tools-extra/clang-tidy/misrac/MisracTidyModule.cpp
@@ -23,6 +23,7 @@
 #include "Rule14p4Check.h"
 #include "Rule15p1Check.h"
 #include "Rule15p2Check.h"
+#include "Rule15p4Check.h"
 #include "Rule15p5Check.h"
 #include "Rule15p6Check.h"
 #include "Rule15p7Check.h"
@@ -89,6 +90,8 @@ public:
         "misrac-rule-15p1");
     CheckFactories.registerCheck<Rule15p2Check>(
         "misrac-rule-15p2");
+    CheckFactories.registerCheck<Rule15p4Check>(
+        "misrac-rule-15p4");
     CheckFactories.registerCheck<Rule15p5Check>(
         "misrac-rule-15p5");
     CheckFactories.registerCheck<Rule15p6Check>(